+++ src/main/kotlin/io/github/ladysnake/chenille/helpers/GithubReleaseHelper.kt
@@ -19,10 +19,9 @@ package io.github.ladysnake.chenille.helpers
 
 import com.github.breadmoirai.githubreleaseplugin.GithubReleaseExtension
 import io.github.ladysnake.chenille.ChenilleProject
-import net.fabricmc.loom.task.RemapJarTask
 
 internal object GithubReleaseHelper {
-    fun configureDefaults(project: ChenilleProject) {
+    fun configureDefaults(project: ChenilleProject, mainArtifact: Any) {
         project.plugins.apply("com.github.breadmoirai.github-release")
 
         project.extensions.configure(GithubReleaseExtension::class.java) {
@@ -33,7 +32,7 @@ internal object GithubReleaseHelper {
             project.git?.run { it.setTargetCommitish { currentBranch() } }
             it.setBody(project.changelog)
 
-            it.setReleaseAssets(project.tasks.named("remapJar", RemapJarTask::class.java))
+            it.setReleaseAssets(mainArtifact)
         }
     }
 }